Adds a segment control to switch the collection view between hourly and daily data

+++ Weather-App/Models/WeatherData.swift
@@ -38,7 +38,7 @@ struct Currently: Codable {
     let windBearing: Int
     let cloudCover: Double
     let uvIndex: Int
-    let visibility: Int
+    let visibility: Double
     let ozone: Double
 }
